build: moving eds cluster to extensions
