+++ CODEOWNERS
@@ -284,6 +284,7 @@ extensions/filters/http/oauth2 @derekargueta @snowp
 /*/extensions/clusters/original_dst @UNOWNED @UNOWNED
 /*/extensions/clusters/logical_dns/ @UNOWNED @UNOWNED
 /*/extensions/clusters/common/ @UNOWNED @UNOWNED
+/*/extensions/clusters/eds/ @UNOWNED @UNOWNED
 
 # URL Pattern Match and Rewrite Library
 /*/extensions/path/uri_template_lib @alyssawilk @yanjunxiang-google
