+++ source/extensions/clusters/eds/eds.cc
@@ -1,4 +1,4 @@
-#include "source/common/upstream/eds.h"
+#include "source/extensions/clusters/eds/eds.h"
 
 #include "envoy/common/exception.h"
 #include "envoy/config/cluster/v3/cluster.pb.h"
