+++ source/extensions/clusters/eds/eds.h
@@ -17,8 +17,8 @@
 
 #include "source/common/config/subscription_base.h"
 #include "source/common/upstream/cluster_factory_impl.h"
-#include "source/common/upstream/leds.h"
 #include "source/common/upstream/upstream_impl.h"
+#include "source/extensions/clusters/eds/leds.h"
 
 namespace Envoy {
 namespace Upstream {
