+++ test/extensions/clusters/eds/eds_speed_test.cc
@@ -15,7 +15,7 @@
 #include "source/common/config/utility.h"
 #include "source/common/config/xds_mux/grpc_mux_impl.h"
 #include "source/common/singleton/manager_impl.h"
-#include "source/common/upstream/eds.h"
+#include "source/extensions/clusters/eds/eds.h"
 #include "source/server/transport_socket_config_impl.h"
 
 #include "test/benchmark/main.h"
