+++ test/extensions/clusters/eds/eds_test.cc
@@ -9,7 +9,7 @@
 
 #include "source/common/config/utility.h"
 #include "source/common/singleton/manager_impl.h"
-#include "source/common/upstream/eds.h"
+#include "source/extensions/clusters/eds/eds.h"
 #include "source/server/transport_socket_config_impl.h"
 
 #include "test/common/stats/stat_test_utility.h"
