+++ tools/code_format/config.yaml
@@ -290,6 +290,7 @@ unsorted_flags:
 # https://github.com/envoyproxy/envoy/issues/9953
 # PLEASE DO NOT ADD FILES TO THIS LIST WITHOUT SENIOR MAINTAINER APPROVAL
 visibility_excludes:
+- source/extensions/clusters/eds/
 - source/extensions/clusters/strict_dns/
 - source/extensions/clusters/static/
 - source/extensions/clusters/original_dst/
